requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
gast==0.2.2
google-auth==1.8.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
liac-arff==2.4.0
lime==0.1.1.36
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
openml==0.10.2
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
pmlb==0.3
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.11.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==223
pywinpty==0.5.5
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.3
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.13.0
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
